feat: introduces serve subcommand for operator logic
