+++ deploy/istio-workspace/operator.yaml
@@ -36,7 +36,7 @@ objects:
             - name: istio-workspace
               image: ${DOCKER_REGISTRY}/${DOCKER_REPOSITORY}/${IMAGE_NAME}:${IMAGE_TAG}
               command:
-              - ike
+              - ike serve
               imagePullPolicy: Always
               env:
                 - name: WATCH_NAMESPACE
